// node-sdk/src/Entities/Enums/TransactionType.ts
export enum TransactionType {
  DECLINE = 1,
  VERIFY = 2,
  CAPTURE = 4,
  AUTH = 8,
  REFUND = 16,
  REVERSAL = 32,
  SALE = 64,
  EDIT = 128,
  VOID = 256,
  ADD_VALUE = 512,
  BALANCE = 1024,
  ACTIVATE = 2048,
  ALIAS = 4096,
  REPLACE = 8192,
  REWARD = 16384,
  DEACTIVATE = 32768,
  BATCH_CLOSE = 65536,
  CREATE = 131072,
  DELETE = 262144,
  FETCH = 524288,
  SEARCH = 1048576,
  HOLD = 2097152,
  RELEASE = 4194304,
  DCC_RATE_LOOKUP = 8388608,
  VERIFY_ENROLLED = 16777216,
  VERIFY_SIGNATURE = 33554432,
  TOKEN_DELETE = 67108864,
  RISK_ASSESS = 134217728,
  INITIATE_AUTHENTICATION = 268435456,
  DATA_COLLECT = 536870912,
  PRE_AUTH_COMPLETION = 1073741824,
  TOKEN_UPDATE = 2147483648,
  BENEFIT_WITHDRAWAL = 4294967296,
  TOKENIZE = 8589934592,
  CREATE_ACCOUNT = 17179869184,
  RESET_PASSWORD = 34359738368,
  RENEW_ACCOUNT = 68719476736,
  UPDATE_OWNERSHIP_DETAILS = 137438953472,
  UPLOAD_CHARGEBACK_DOCUMENT = 274877906944,
  UPLOAD_DOCUMENT = 549755813888,
  OBTAIN_SSO_KEY = 1099511627776,
  UPDATE_BANK_ACCOUNT_OWNERSHIP = 2199023255552,
  ADD_FUNDS = 4398046511104,
  SWEEP_FUNDS = 8796093022208,
  ADD_CARD_FLASH_FUNDS = 17592186044416,
  PUSH_MONEY_FLASH_FUNDS = 35184372088832,
  DISBURSE_FUNDS = 70368744177664,
  SPEND_BACK = 140737488355328,
  REVERSE_SPLITPAY = 281474976710656,
  SPLIT_FUNDS = 562949953421312,
  GET_ACCOUNT_DETAILS = 1125899906842624,
  GET_ACCOUNT_BALANCE = 2251799813685248,
  DETOKENIZE = 4503599627370496,
  DISPUTE_ACCEPTANCE = 9007199254740992,
  DISPUTE_CHALLENGE = 18014398509481984,
  REAUTH = 36028797018963968,
  CONFIRM = 72057594037927936,
  PAYBYLINK_UPDATE = 144115188075855872,
  TRANSFER_FUNDS = 288230376151711744,
  DEVICE_ORDER = 576460752303423488,
  GET_TOKEN_INFO = 4461637978905770969,
  HOSTED_PAYMENT_PAGE = 8935141660703064064,
}